+++ CHANGELOG.md
@@ -10,6 +10,8 @@
   - Adding field `jobs_max`: Number of maximal jobs that can be run in parallel
   - Adding field `queue_max`: Number of maximal queue size
 - Update Python dependencies
+- Automatically build `linux/amd64` and `linux/arm64` variants of the Docker image
+- Update GitHub actions workflow for Docker image build and release
 
 
 ## Version 3.3.9 (2026-02-05)